requirements.txt
pygithub~=1.54
pyyaml~=5.4
click~=8.0
packaging~=21.0
cloudshell-pip-download~=0.5
attrs~=21.0
